refactor: function settings
