+++ src/routes/console/project-[project]/functions/function-[function]/settings/deleteModal.svelte
